MULE-16636: CursorProvider is retained too long when cursors are opened but not consumed
