+++ core/src/main/java/org/mule/runtime/core/internal/streaming/CursorProviderJanitor.java
@@ -89,6 +89,9 @@ public class CursorProviderJanitor {
   public void releaseCursor(Cursor cursor) {
     try {
       cursor.release();
+      if (cursors.remove(cursor) && provider.isClosed() && cursors.isEmpty()) {
+        releaseResources();
+      }
     } catch (Exception e) {
       LOGGER.warn("Exception was found trying to release cursor resources. Execution will continue", e);
     } finally {
